% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVolcano.R
\name{add_regulate}
\alias{add_regulate}
\title{Add regulate column to the DEG result}
\usage{
add_regulate(
  data,
  log2FC_name = "log2FoldChange",
  fdr_name = "padj",
  log2FC = 1,
  fdr = 0.05
)
}
\arguments{
\item{data}{a data of DEG result.}

\item{log2FC_name}{the column name corresponding to log2FC}

\item{fdr_name}{the column name corresponding to fdr(p-adjust)}

\item{log2FC}{the threshold value of log2FC.}

\item{fdr}{the threshold value of fdr.}
}
\value{

}
\description{
Add regulate column to the DEG result
}
\examples{
# load the data
data(deg_data)
head(deg_data)

data_new <- add_regulate(deg_data, log2FC_name = "log2FoldChange",
                         fdr_name = "padj",log2FC = 1, fdr = 0.05)
head(data_new)
}
