% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVolcano-package.R
\docType{package}
\name{ggVolcano-package}
\alias{ggVolcano}
\alias{ggVolcano-package}
\title{ggVolcano: Make a nice volcano plot}
\description{
You can make a beautiful volcano plot with just one line of code.
}
\keyword{internal}
