% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVolcano.R
\name{gradual_volcano}
\alias{gradual_volcano}
\title{A gradual volcano plot}
\usage{
gradual_volcano(
  data,
  x = "log2FoldChange",
  y = "padj",
  pointSizeRange = c(0.5, 4),
  fills = c("#39489f", "#39bbec", "#f9ed36", "#f38466", "#b81f25"),
  colors = c("#17194e", "#68bfe7", "#f9ed36", "#a22f27", "#211f1f"),
  x_lab = NULL,
  y_lab = NULL,
  legend_title = NULL,
  add_line = TRUE,
  label = "row",
  label_number = 10,
  output = TRUE,
  filename = "volcano_plot"
)
}
\arguments{
\item{data}{a data of DEG result.}

\item{x}{the column name corresponding to the x-axis, defualt is "log2FoldChange"}

\item{y}{the column name corresponding to the y-axis, defualt is "padj".}

\item{pointSizeRange}{a two elements vector specifing the range of the point size.}

\item{fills}{a vector containing the fill color of the point.}

\item{colors}{a vector containing the stroke color of the point.}

\item{x_lab}{label of x-axis.}

\item{y_lab}{label of y-axis.}

\item{legend_title}{title of the legend.}

\item{add_line}{a logical value, whether to add a dashed line, defult is TRUE.}

\item{label}{the column name corresponding to the label.}

\item{label_number}{how many gene labels you want to show in the plot.}

\item{output}{a logical value, whether to save the image, defult is TRUE.}

\item{filename}{if the output = TRUE, please set a filename.}
}
\value{

}
\description{
A gradual volcano plot
}
\examples{
# load the data
data(deg_data)
data <- add_regulate(deg_data, log2FC_name = "log2FoldChange",
                     fdr_name = "padj",log2FC = 1, fdr = 0.05)

# plot
gradual_volcano(deg_data, x = "log2FoldChange", y = "padj",
             label = "row", label_number = 10, output = FALSE)
}
