% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVolcano.R
\name{term_volcano}
\alias{term_volcano}
\title{A term volcano plot}
\usage{
term_volcano(
  data,
  term_data,
  x = "log2FoldChange",
  y = "padj",
  normal_point_color = "#999999",
  normal_point_size = 1,
  deg_point_color = "black",
  deg_point_fill = c(dendritic = "#49c2c6", `ion transport.` = "#fbcbcc", metabolic =
    "#eef0ac", myelin = "#b1daa7", synaptic = "#d0d0a0"),
  deg_point_size = 2,
  legend_background_fill = "#fefde2",
  legend_title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  add_line = TRUE,
  label = "row",
  label_number = 10,
  output = TRUE,
  filename = "volcano_plot"
)
}
\arguments{
\item{data}{a data of DEG result.}

\item{term_data}{a two columns dataframe containing some genes' GO terms information.}

\item{x}{the column name corresponding to the x-axis, defualt is "log2FoldChange"}

\item{y}{the column name corresponding to the y-axis, defualt is "padj".}

\item{normal_point_color}{the color of normal points. Defult is "#999999".}

\item{normal_point_size}{the size of normal points. Defult is 1.}

\item{deg_point_color}{the stroke color of these deg points, defualt is "black".}

\item{deg_point_fill}{a named vector containing the fill color of these deg points. Make sure that the names of the elements in the vector correspond to the pathways and colors.}

\item{deg_point_size}{the size of these deg points. Defult is 2.}

\item{legend_background_fill}{the fill color of legend background.}

\item{legend_title}{title of the legend.}

\item{x_lab}{label of x-axis.}

\item{y_lab}{label of y-axis.}

\item{add_line}{a logical value, whether to add a dashed line, defult is TRUE.}

\item{label}{the column name corresponding to the label.}

\item{label_number}{how many gene labels you want to show in the plot.}

\item{output}{a logical value, whether to save the image, defult is TRUE.}

\item{filename}{if the output = TRUE, please set a filename.}
}
\value{

}
\description{
A term volcano plot
}
\examples{
# load the data
data(deg_data)
data(term_data)
data <- add_regulate(deg_data, log2FC_name = "log2FoldChange",
                     fdr_name = "padj",log2FC = 1, fdr = 0.05)

# plot
term_volcano(deg_data, term_data, x = "log2FoldChange", y = "padj",
             label = "row", label_number = 10, output = FALSE)
}
