% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVolcano.R
\name{ggvolcano}
\alias{ggvolcano}
\title{A volcano plot}
\usage{
ggvolcano(
  data,
  x = "log2FoldChange",
  y = "padj",
  pointSize = 1,
  pointShape = 21,
  fills = c("#00AFBB", "#999999", "#FC4E07"),
  colors = c("#00AFBB", "#999999", "#FC4E07"),
  x_lab = NULL,
  y_lab = NULL,
  legend_title = NULL,
  legend_position = "UL",
  log2FC_cut = 1,
  FDR_cut = 0.05,
  add_line = TRUE,
  add_label = TRUE,
  label = "row",
  label_number = 10,
  custom_label = NULL,
  output = TRUE,
  filename = "volcano_plot"
)
}
\arguments{
\item{data}{a data of DEG result.}

\item{x}{the column name corresponding to the x-axis, defualt is "log2FoldChange"}

\item{y}{the column name corresponding to the y-axis, defualt is "padj".}

\item{pointSize}{size of the point.}

\item{pointShape}{shape of the point.}

\item{fills}{a vector containing the fill color of the point.}

\item{colors}{a vector containing the stroke color of the point.}

\item{x_lab}{label of x-axis.}

\item{y_lab}{label of y-axis.}

\item{legend_title}{title of the legend.}

\item{legend_position}{the position of legend. You can choose one from "UL" -- Up Left, "UR" -- Up Right, "DL" -- Down Left, "DR" -- Down Right}

\item{log2FC_cut}{cutoff value of log2FC.}

\item{FDR_cut}{cutoff value of FDR.}

\item{add_line}{a logical value, whether to add a dashed line, defult is TRUE.}

\item{add_label}{a logical value, whether to add gene label, defult is TRUE.}

\item{label}{the column name corresponding to the label.}

\item{label_number}{how many gene labels you want to show in the plot.}

\item{custom_label}{a vector containing your interest gene names that you want to add to the plot.}

\item{output}{a logical value, whether to save the image, defult is TRUE.}

\item{filename}{if the output = TRUE, please set a filename.}
}
\value{

}
\description{
A volcano plot
}
\examples{
# load the data
data(deg_data)
data <- add_regulate(deg_data, log2FC_name = "log2FoldChange",
                     fdr_name = "padj",log2FC = 1, fdr = 0.05)
# plot
ggvolcano(data, x = "log2FoldChange", y = "padj",
          label = "row", label_number = 10, output = FALSE)
}
